feat: middleware admin, customer

+++ route.php
@@ -3,6 +3,8 @@
 
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\ProductController;
+use App\Http\Middlewares\AdminAuthenticate;
+use App\Http\Middlewares\CustomerAuthenticate;
 use Libraries\Redirect\Route;
 
 Route::post('/register', [AuthController::class, 'register']);
@@ -10,6 +12,6 @@ Route::post('/login', [AuthController::class, 'login']);
 
 Route::get('/product', [ProductController::class, 'index']);
 Route::get('/product/{id}', [ProductController::class, 'show']);
-Route::post('/product', [ProductController::class, 'store']);
-Route::put('/product/{id}', [ProductController::class, 'update']);
-Route::delete('/product/{id}', [ProductController::class, 'destroy']);
+Route::post('/product', [ProductController::class, 'store'])->middleware(AdminAuthenticate::class);
+Route::put('/product/{id}', [ProductController::class, 'update'])->middleware(AdminAuthenticate::class);
+Route::delete('/product/{id}', [ProductController::class, 'destroy'])->middleware(AdminAuthenticate::class);
